Add support for downloading binary files using S3Client.getObject
This commit adds an `additionalHeaders` parameter to S3Client.getObject.
The operation additionally receives support for the `Accept` header.
When the header is provided with the value `application/octet-stream`
the operation will, as a result, treat the response as binary content
and return an ArrayBuffer.

We decided to opt for the header approach, as it doesn't leak nor depend
on k6 interfaces, and adopts a standard HTTP behavior, indicating to
the server the type of content we expect in return, through the Accept
header.