Disregard `K6_DISCARD_RESPONSE_BODIES` (#114)
* Add baseRequestParams to AWSClient

This allow us to ignore k6's discardResponseBodies in the context of
jslib-aws. As it is mostly a utility library, users have been mostly
about having the ability available. See #45 for details.

Using the baseRequestParams we set the default responseType for children
implementation, allowing us to effectively disregard the
http discardResponseBodies option. We can still, locally override it
where necessary, as in S3Client.getObject for instance.

* Adopt AWSClient.baseRequestParams in AWS client classes

* Update src/internal/client.ts

Co-authored-by: Joan López de la Franca Beltran <5459617+joanlopez@users.noreply.github.com>

---------

Co-authored-by: Joan López de la Franca Beltran <5459617+joanlopez@users.noreply.github.com>